将controller,service,mapper和entity移入NucTest包

+++ src/main/java/com/segroup/hospitalsite/NucTest/controller/NucTestAppController.java
@@ -1,8 +1,8 @@
 package com.segroup.hospitalsite.NucTest.controller;
-
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.segroup.hospitalsite.NucTest.service.INucTestappService;
-import com.segroup.hospitalsite.NucTest.entity.NucTestapp;
+import com.segroup.hospitalsite.NucTest.entity.NucTestApp;
+import com.segroup.hospitalsite.UserInfo.utils.JsonResult;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.stereotype.Controller;
 
 import java.time.LocalDate;
+import java.time.format.DateTimeParseException;
 import java.util.List;
 
 /**
@@ -25,7 +26,16 @@ import java.util.List;
 @RequestMapping("/nucTestApp")
 @Controller
 @Api(tags = "核酸检测数据接口")
-public class NucTestappController {
+public class NucTestAppController{
+
+    public static final int SUCCESS = 200;
+    public static final int ID_NOT_FOUND_ERROR = 4002;
+    public static final int DATE_ILLEGAL_ERROR = 4003;
+    public static final int REMOVE_ERROR = 4004;
+    public static final int INSERT_ERROR = 4005;
+    public static final int NULL_ERROR = 4006;
+    public static final int INSERTION_ERROR = 5000;
+    public static final int UPDATE_ERROR = 5001;
     @Autowired
     private INucTestappService ntaService;
 
@@ -40,6 +50,7 @@ public class NucTestappController {
     public String Query(){
         return "query"; //需要换成实际的html页面，放到templates/中
     }
+
     @ApiOperation(value = "修改核酸预约日期", notes = "根据身份证号修改")
     @ApiImplicitParams({
             @ApiImplicitParam(paramType = "path", name = "usr_id", value = "身份证号", required = true),
@@ -47,36 +58,70 @@ public class NucTestappController {
     })
     @GetMapping("/update/{usr_id}/{new_date}")
     @ResponseBody
-    public String updateRecord(@PathVariable String usr_id, @PathVariable String new_date){
-        LocalDate t_new_date = LocalDate.parse(new_date);
+    public JsonResult<Void> updateRecord(@PathVariable String usr_id, @PathVariable String new_date){
+        JsonResult<Void> result;
+        QueryWrapper<NucTestApp> wrapper = new QueryWrapper<>();
+        wrapper.eq("usr_id", usr_id);
+        List queryList = ntaService.list(wrapper);
+        LocalDate t_new_date;
+        try{
+            t_new_date = LocalDate.parse(new_date);
+        }
+        catch (DateTimeParseException e)
+        {
+            result = new JsonResult<Void>(DATE_ILLEGAL_ERROR, "非法的日期格式，应为YYYY-mm-dd");
+            return result;
+        }
+        if(queryList.size() == 0){
+            result = new JsonResult<Void>(ID_NOT_FOUND_ERROR, "该用户没有核酸检测预约记录");
+        }
+        else{
+            result = new JsonResult<Void>(SUCCESS, "修改预约日期至: "+new_date);
+        }
         ntaService.updateByUsrId(usr_id, t_new_date);
-        return "成功修改预约日期为: "+ new_date;
+        return result;
     }
 
     @ApiOperation(value = "取消核酸检测预约", notes = "根据身份证号取消核酸检测预约")
     @ApiImplicitParam(paramType = "path", name = "usr_id", value="身份证号", required = true)
     @GetMapping("delete/{usr_id}")
     @ResponseBody
-    public String deleteRecord(@PathVariable String usr_id){
-        QueryWrapper<NucTestapp> wrapper = new QueryWrapper<>();
+    public JsonResult<Void>  deleteRecord(@PathVariable String usr_id){
+        JsonResult<Void> result;
+        QueryWrapper<NucTestApp> wrapper = new QueryWrapper<>();
         wrapper.eq("usr_id", usr_id);
+        List queryList = ntaService.list(wrapper);
+        if(queryList.size() == 0){
+            result = new JsonResult<>(ID_NOT_FOUND_ERROR, "该用户没有核酸检测预约记录");
+            return result;
+        }
         boolean flag = ntaService.remove(wrapper);
         if(flag){
-            return "取消成功";
+            result = new JsonResult<>(REMOVE_ERROR, "取消预约成功！");
         }
-        else {
-            return "取消失败";
+        else{
+            result = new JsonResult<>(SUCCESS, "删除记录失败！");
         }
+        return result;
     }
 
     @ApiOperation(value = "核酸检测预约记录查询", notes = "根据身份证号查询核酸检测预约记录")
     @ApiImplicitParam(paramType = "path", name = "usr_id", value="身份证号", required = true)
     @GetMapping ("/query/{usr_id}")
     @ResponseBody
-    public List<NucTestapp> queryRecord(@PathVariable String usr_id){
-        QueryWrapper<NucTestapp> wrapper = new QueryWrapper<>();
+    public JsonResult<NucTestApp> queryRecord(@PathVariable String usr_id){
+        JsonResult<NucTestApp> result;
+        QueryWrapper<NucTestApp> wrapper = new QueryWrapper<>();
         wrapper.eq("usr_id", usr_id);
-        return ntaService.list(wrapper);
+        List<NucTestApp> queryList = ntaService.list(wrapper);
+        if(queryList.size() == 0){
+            result = new JsonResult<>(ID_NOT_FOUND_ERROR, "该用户没有核酸检测预约记录");
+        }
+        else{
+            result = new JsonResult<>(SUCCESS, "查询成功，返回用户预约记录");
+            result.setData(queryList.get(0));
+        }
+        return result;
     }
 
     @ApiOperation(value = "核酸预约记录新增", notes = "根据提交的预约信息增加记录")
@@ -88,17 +133,25 @@ public class NucTestappController {
     })
     @PostMapping ("/insert")
     @ResponseBody
-    public String insertAppRecord(@RequestParam String usrName,
+    public JsonResult<NucTestApp> insertAppRecord(@RequestParam String usrName,
                                   @RequestParam String usrId,
                                   @RequestParam Integer testType,
                                   @RequestParam String testDate){
-        NucTestapp ntaInfo = new NucTestapp();
+        JsonResult<NucTestApp> result;
+        NucTestApp ntaInfo = new NucTestApp();
         ntaInfo.setUsrName(usrName);
         ntaInfo.setUsrId(usrId);
         ntaInfo.setTestType(testType);
         ntaInfo.setTestDate(LocalDate.parse(testDate));
-        ntaService.save(ntaInfo);
-        return "success";
+        boolean flag = ntaService.save(ntaInfo);
+        if(flag){
+            result = new JsonResult<>(SUCCESS, "预约成功,返回预约信息");
+            result.setData(ntaInfo);
+        }
+        else{
+            result = new JsonResult<>(INSERT_ERROR, "插入记录失败！");
+        }
+        return result;
     }
 
 }
